Added more move actions
Added more move action web sockets and buttons which sends the data to
and from the server. Note still need to check the JSON i am sending for
each one as it is not correct at the moment. Once back end logic is all
ready these functions can be used to glue the logic to the web sockets
and get the info going to and from the client