UI improved look
The user interface was improved and its now divided in sections. Home
page, new game page and join game page.
Home page :
This is where you can select to see the rules, start a new game or join
a existing game. To access the home page call http://127.0.0.1:5000/home
Home page :
This is where you can select to see the rules, start a new game or join
a existing game. To access the home page call http://127.0.0.1:5000/home
Home page :
This is where you can select to see the rules, start a new game or join
a existing game. To access the home page call http://127.0.0.1:5000/home
Home page :
This is where you can select to see the rules, start a new game or join
a existing game. To access the home page call http://127.0.0.1:5000/home