Glued together functions
Got the back end now working with the front end for end of round
infections. Still need to work out the path for outbreaks. Also have
dispatcher player action working and implemented check for medic to
make sure he treats cities automatically when he moves if there is a
cure