Added in onLoad features and more
Now get the cube levels, outbreak and infection levels on page load
instead of just at the end of rounds. Also added in more checking in
the lobby so that users cannot start a game until everyone in the lobby
has selected a player role. Both front and back end checks to check
this as sometimes if a user has already locked in and then someone
joins, when they then press play it goes ahead when it shouldn’t. So
added in a back end check too.